Create height_of_tree.rs
In this implementation, we define a 'TreeNode' struct that represents a node in a binary tree. The 'TreeNode' struct has a 'height' method that recursively computes the height of the node by computing the maximum height of its left and right children, and adding 1 to account for the current node.

In the 'main' function, we create a binary tree with some sample data, and then call the 'height' method on the root node to compute the height of the tree. Finally, we print out the height of the tree to the console.